docker and deployment configurations

+++ Backend/DTOs/Item/RentalOptionDto.cs
@@ -4,6 +4,6 @@
     {
         public Guid? Id { get; set; }
         public string RentalOptionName { get; set; } = string.Empty;
-        public decimal Price { get; set; } = 0;
+        public int Price { get; set; } = 0;
     }
 }
